requirements.txt
# Core dependencies
# Install: pip install -r requirements.txt

python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
openpyxl>=3.1.0        # Excel (.xlsx) support
